requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.6
django-crispy-forms==1.12.0
django-taggit==1.5.1
filelock==3.0.12
importlib-metadata==4.6.4
Pillow==8.3.1
platformdirs==2.2.0
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.7.2
whitenoise==5.3.0
zipp==3.5.0
